Struct bigquery2::Bigquery

pub struct Bigquery<C, NC, A> {
    // some fields omitted
}

Central instance to access all Bigquery related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate bigquery2;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and `client_secret`, 
// among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about what's going on
// You probably want to bring in your own `TokenStorage` to persist tokens and retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Bigquery::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: Table = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.tables().update(&req, "projectId", "datasetId", "tableId")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Methods

impl<'a, C, NC, A> Bigquery<C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>> + 'a, A: GetToken

fn new(client: C, authenticator: A) -> Bigquery<C, NC, A>

fn datasets(&'a self) -> DatasetMethodsBuilder<'a, C, NC, A>

fn jobs(&'a self) -> JobMethodsBuilder<'a, C, NC, A>

fn projects(&'a self) -> ProjectMethodsBuilder<'a, C, NC, A>

fn tabledata(&'a self) -> TabledataMethodsBuilder<'a, C, NC, A>

fn tables(&'a self) -> TableMethodsBuilder<'a, C, NC, A>